Handle LDAP Socket Connection Errors
Log and handle LDAPSocketOpenError exceptions raised
by the connect_ldap() function.

This will write out a log with the exception class
name followed by the stringified version of the
exception, as LDAP exceptions do not contain
properties for getting at the pertinent text data
detailing what caused the socket issue.

It will also raise an Error() instance with a "user
friendly" text message to be displayed in the web
interface.